# api/openapi.yaml
openapi: 3.0.3
info:
  title: ChessMgr API
  description: Chess Tournament Manager API
  version: 1.0.0
servers:
  - url: /api
security: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Tournament:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        organizer_id:
          type: string
          format: uuid
        pairing_system:
          type: string
          enum: [round_robin, knockout, swiss, scheveningen, manual]
        status:
          type: string
          enum: [draft, registration, active, completed]
        rounds_count:
          type: integer
        time_control:
          type: string
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - organizer_id
        - pairing_system
        - status
        - created_at
        - updated_at
...
    UpdateTournamentRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        pairing_system:
          type: string
          enum: [round_robin, knockout, swiss, scheveningen, manual]
        rounds_count:
          type: integer
        time_control:
          type: string
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        status:
          type: string
          enum: [draft, registration, active, completed]
...
paths:
...
  /tournaments/{id}:
    get:
      summary: Get tournament
      tags: [tournaments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tournament
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tournament'
    put:
      summary: Update tournament
      tags: [tournaments]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTournamentRequest'
      responses:
        '200':
          description: Updated tournament
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tournament'
    delete:
      summary: Delete tournament
      tags: [tournaments]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
...